adicionando view e controller

+++ sistemaVendas/src/main/kotlin/model/SaleProducts.kt
@@ -1,5 +1,5 @@
 package model
 
 class SaleProducts (val saleId: Long, val productId: Long){
-    var quantity:Long = 0
+    var quantity:Int = 0
 }
